# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  sales.CreateCustomerRequest:
    properties:
      email:
        example: johndoe@example.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
  sales.CreateOrderRequest:
    properties:
      customer_id:
        example: a4b6c8d0-1234-5678-90ab-cdef12345678
        type: string
      items:
        items:
          $ref: '#/definitions/sales.OrderItemRequest'
        type: array
    type: object
  sales.CustomerResponse:
    properties:
      customer_id:
        example: a4b6c8d0-1234-5678-90ab-cdef12345678
        type: string
      email:
        example: johndoe@example.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
  sales.OrderItemRequest:
    properties:
      product_id:
        example: bc731ee0-3333-4f9e-a89c-a1d257211140
        type: string
      quantity:
        example: 2
        type: integer
      unit_price:
        description: Client-supplied; no pricing catalog yet
        example: 25000
        type: number
      warehouse_id:
        example: 8f076135-d858-45ec-b9cc-0320df4ee99c
        type: string
    type: object
  sales.OrderItemResponse:
    properties:
      product_id:
        example: bc731ee0-3333-4f9e-a89c-a1d257211140
        type: string
      quantity:
        example: 2
        type: integer
      unit_price:
        example: 25000
        type: number
      warehouse_id:
        example: 8f076135-d858-45ec-b9cc-0320df4ee99c
        type: string
    type: object
  sales.OrderResponse:
    properties:
      customer_id:
        example: a4b6c8d0-1234-5678-90ab-cdef12345678
        type: string
      items:
        items:
          $ref: '#/definitions/sales.OrderItemResponse'
        type: array
      order_id:
        example: d3b07384-d113-4956-953e-52f01f05e3d9
        type: string
      status:
        description: e.g., pending, fulfilled, stock_issue
        example: fulfilled
        type: string
      total:
        example: 50000
        type: number
    type: object
  tenancy.TenantResponse:
    properties:
      created_at:
        type: string
      name:
        type: string
      tenant_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant REP system for OneDesk.
  title: OneDesk API
  version: "1.0"
paths:
...
  /tenants/me:
    get:
      description: Retrieves structural information about the authenticated user's
        tenant account context.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tenancy.TenantResponse'
        "401":
          description: missing tenant context / unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current tenant profile
      tags:
      - tenancy
  /v1/customers:
    get:
      description: Returns all customer profiles matching the active tenant organizational
        boundary.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sales.CustomerResponse'
            type: array
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal service database lookup fault
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List tenant customer accounts
      tags:
      - sales
    post:
      consumes:
      - application/json
      description: Registers a new customer account record attached to the current
        active tenant workspace.
      parameters:
      - description: Customer details input parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sales.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sales.CustomerResponse'
        "400":
          description: Invalid payload data syntax or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new customer profile
      tags:
      - sales
...
  /v1/orders:
    post:
      consumes:
      - application/json
      description: Submits a sales checkout request, consumes monthly plan order entitlement
        quotas, and runs transactional inventory stock deductions.
      parameters:
      - description: Order checklist containing items list details mapping
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sales.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sales.OrderResponse'
        "400":
          description: Invalid references or schema entity fields input data validation
            errors
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Active monthly plan order transaction cap limit exhausted
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Order registered but partial stock depletion triggered a fallback
            state
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Place a new sales order
      tags:
      - sales
  /v1/orders/{orderId}:
    get:
      description: Fetches the comprehensive order composition manifest, settlement
        pricing math summaries, and current routing lifecycle status tags.
      parameters:
      - description: System transaction reference token lookup ID (UUID)
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sales.OrderResponse'
        "400":
          description: Malformed track sequence path identifier value format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No transactional invoice matches the specified trace index
            ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Core data repository extraction execution block runtime failure
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get details of a sales order
      tags:
      - sales
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"